CheeseWindow: Connect to the correct CheeseCamera signal.
* src/cheese-window.vala: Connect to state-flags-changed, not
state-changed, which does not exist and which caused a runtime warning.
* src/vapi/cheese-common.vapi: Correct the signal name here by hand,
because I do not know how to properly regenerate this file.
Presumably the generator used the (previously incorrect) default
signal handler name to guess the name of the signal.

For me, this allows us to actually respond when the camera device
becomes ready, so I can actually take a picture.