CheeseCamera: Avoid crash when gst_*_message_parse() fails.
* libcheese/cheese-camera.c: cheese_camera_bus_message_cb():
Sometimes the GstMessage's internal structure seems to not be for a
warning message, even when GST_MESSAGE_TYPE() says that it is.
We should generally check the err before dereferencing it anyway,
and that prevents this from crashing cheese.
This happens when stopping video recording. The video is saved anyway.